Created Report and Form Component
