+++ .env.example
@@ -1,0 +1,1 @@
+VITE_GEMINI_API_KEY=AIzaSyCmKygcasVxAkUt_I6yPKQIXA88h45KQxI
